backend/requirements.txt
google-generativeai
python-dotenv
fastapi
uvicorn[standard]
qdrant-client
sentence-transformers
torch
transformers